Revert "crypto, quiet key input"

See merge request zt40/storage-project!11

+++ src/utils.hpp
@@ -106,7 +106,7 @@ std::vector<Meta> readAllMeta(char* buffer, int size);
 TreeNode* generateTree(std::vector<Meta> metaList);
 
 // For generate Image
-int writeImageMeta(std::string name, long start, long size, unsigned int permission, unsigned int owner, unsigned int group, bool isDir, time_t lastModified, FILE* file);
+int writeImageMeta(std::string name, size_t start, size_t size, unsigned int permission, unsigned int owner, unsigned int group, bool isDir, time_t lastModified, FILE* file);
 
 int writeAllMeta(std::vector<Meta> metas, FILE* file);
 
@@ -116,11 +116,11 @@ int generateImage(const std::string& directory, const std::string& image);
 // int ecb_decrypt(unsigned char* in, int in_len, unsigned char* out, unsigned char* key);
 
 int encrypt(const std::string& path, const std::string& key);
-int decrypt(FILE* f, unsigned char* keyhash, int st_blk, int ed_blk, unsigned char* buffer);
+int decrypt(FILE* f, unsigned char* keyhash, size_t st_blk, size_t ed_blk, unsigned char* buffer);
 
 // int getImageSize(const std::string& path);
 
-int readEncImage(FILE* f, unsigned char* keyhash, unsigned char* buffer, int begin, int len);
+size_t readEncImage(FILE* f, unsigned char* keyhash, unsigned char* buffer, int begin, int len);
 std::vector<Meta> readEncMeta(FILE* f, unsigned char* keyhash);
 
 int verify(FILE* f, unsigned char* keyhash);
